+++ vcls-invoicing/__manifest__.py
@@ -15,7 +15,7 @@
     # Check https://github.com/odoo/odoo/blob/12.0/odoo/addons/base/data/ir_module_category_data.xml
     # for the full list
     'category': 'Uncategorized',
-    'version': '1.7.62',
+    'version': '1.7.61',
 
 
     # any module necessary for this one to work correctly
