+++ vcls_subscription/__manifest__.py
@@ -11,7 +11,7 @@
     'author': "VCLS",
     'website': "http://www.voisinconsulting.com",
     'category': 'Uncategorized',
-    'version': '0.0.51',
+    'version': '0.0.52',
     'depends': [
         'base',
         'sale_subscription',
