+++ vcls_subscription/models/sale_subscription.py
@@ -11,7 +11,6 @@ from uuid import uuid4
 from odoo import api, fields, models, _
 from odoo.exceptions import UserError, ValidationError
 from odoo.osv import expression
-from odoo.tools import format_date
 from odoo.tools.safe_eval import safe_eval
 
 from odoo.addons import decimal_precision as dp
@@ -119,3 +118,35 @@ class SaleSubscription(models.Model):
             to_update.write({
                 'qty_delivered_method':'manual',
             })
+
+class SaleOrderLine(models.Model):
+    _inherit = 'sale.order.line'
+
+    def _prepare_invoice_line(self, qty):
+        """
+        Override to add subscription-specific behaviors.
+
+        Display the invoicing period in the invoice line description, link the invoice line to the
+        correct subscription and to the subscription's analytic account if present.
+        """
+        res = super(SaleOrderLine, self)._prepare_invoice_line(qty)
+        temp_name = res['name'].split('\nInvoicing period: ')[0]
+        res.update(name=temp_name)
+        if self.subscription_id:
+            res.update(subscription_id=self.subscription_id.id)
+            # In VCLS we invoice the previous period
+            if self.order_id.subscription_management != 'upsell':
+                periods = {'daily': 'days', 'weekly': 'weeks', 'monthly': 'months', 'yearly': 'years'}
+                next_date = fields.Date.from_string(self.subscription_id.recurring_next_date) - relativedelta(**{periods[self.subscription_id.recurring_rule_type]: self.subscription_id.recurring_interval}) + relativedelta(months=+1)
+                previous_date = next_date - relativedelta(**{periods[self.subscription_id.recurring_rule_type]: self.subscription_id.recurring_interval})
+                lang = self.order_id.partner_invoice_id.lang
+                format_date = self.env['ir.qweb.field.date'].with_context(lang=lang).value_to_html
+
+                # Ugly workaround to display the description in the correct language
+                if lang:
+                    self = self.with_context(lang=lang)
+                period_msg = _("Invoicing period: %s - %s") % (format_date(fields.Date.to_string(previous_date), {}), format_date(fields.Date.to_string(next_date - relativedelta(days=1)), {}))
+                res.update(name=res['name'] + '\n' + period_msg)
+            if self.subscription_id.analytic_account_id:
+                res['account_analytic_id'] = self.subscription_id.analytic_account_id.id
+        return res
